fix: PR #4 round 2 — atomic claim, callback-pending, truncated error, int64 CIK, orchestrator guard
- store.go: add ClaimQueuedJob — Firestore transaction CAS that verifies
  status==QUEUED && active_agent==agent before transitioning to IN_PROGRESS.
  Eliminates the read-then-write race in duplicate Cloud Tasks deliveries.

- handlers.go: replace GetJob+guard+UpdateJob with a single ClaimQueuedJob
  call for race-safe duplicate delivery protection.

- handlers.go: before enqueueing the orchestrator callback, transition job
  to QUEUED+orchestrator so a future sweep can recover stranded jobs if
  the enqueue fails.

- handlers.go: include truncated (≤200 char) error detail in final_result
  when an agent fails, instead of just the agent name.

- data_agent.go: change CIK field from int to int64 to prevent overflow
  on 32-bit platforms.

- orchestrator.go: add active_agent==orchestrator guard before the circuit
  breaker, preventing duplicate /internal/route deliveries from bumping
  hop_count or overwriting an in-flight agent.

+++ orchestrator/internal/data_agent.go
@@ -123,7 +123,7 @@ func (c *cikCache) populate(ctx context.Context, httpClient *http.Client, userAg
 
 	// Response format: {"0": {"cik_str": 320193, "ticker": "AAPL", "title": "Apple Inc"}, ...}
 	var entries map[string]struct {
-		CIK    int    `json:"cik_str"`
+		CIK    int64  `json:"cik_str"`
 		Ticker string `json:"ticker"`
 		Title  string `json:"title"`
 	}
